requirements.txt
# Core dependencies
click>=8.1.0
clickhouse-connect>=0.6.0
ruamel.yaml>=0.17.0
sqlparse>=0.4.0
rich>=13.0.0
pydantic>=2.0.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
